moved factory code out of app
